+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Stable sorting of tasks (#371)
 
+### Added
+- Expose the orchestrator version and chaincode version (#370)
+
 ## [0.0.2] - 2021-09-16
 
 ### Added